requirements.txt
Django==4.0.4
django-import-export==2.8.0
djangorestframework==3.13.1
gunicorn==20.1.0
whitenoise==5.3.0